% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erikdata_DATA.R
\docType{data}
\name{dat_atus}
\alias{dat_atus}
\title{USA GLS.gov ATUS 2003-2021 Multi-Year Microdata Files}
\format{
tibble
}
\source{
\code{https://www.bls.gov/tus/data/datafiles-0321.htm}
}
\usage{
data(dat_atus)
}
\description{
Joined Respondent and Activity summary files; with aggregated categories; version Erhardt-202212
}
\examples{
data(dat_atus)
str(dat_atus)
}
\keyword{datasets}
