% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erikdata_logo.R
\name{erikdata_logo}
\alias{erikdata_logo}
\title{The erikdata logo, using ASCII or Unicode characters}
\usage{
erikdata_logo(sw_unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{sw_unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\description{
Use \code{\link[crayon:strip_style]{crayon::strip_style()}} to get rid of the colors.
}
\examples{
erikdata_logo()
}
