% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{erikdata_packages}
\alias{erikdata_packages}
\title{List all packages in the erikdata}
\usage{
erikdata_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include erikdata in the list?}
}
\description{
List all packages in the erikdata
}
\examples{
erikdata_packages()
}
