% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erikdata_DATA.R
\docType{data}
\name{labels_dat_atus}
\alias{labels_dat_atus}
\title{USA GLS.gov ATUS 2003-2021 Multi-Year Microdata Files}
\format{
tibble
}
\source{
\code{https://www.bls.gov/tus/data/datafiles-0321.htm}
}
\usage{
data(labels_dat_atus)
}
\description{
Variable labels for \code{dat_atus}
}
\examples{
data(labels_dat_atus)
str(labels_dat_atus)
}
\keyword{datasets}
