% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erikdata_logo.R
\name{print.erikdata_logo}
\alias{print.erikdata_logo}
\title{Print for erikdata_logo}
\usage{
\method{print}{erikdata_logo}(x, ...)
}
\arguments{
\item{x}{the \code{erikdata_logo()}}

\item{...}{additional parameters passed to \code{cat()}}
}
\value{
the logo, invisibly
}
\description{
Print for erikdata_logo
}
